Split main.rs
